Fix memory leak bug

+++ app/src/main/java/dev/tanakornsss/luminality/data/backup/BackupRepository.kt
@@ -1,5 +1,7 @@
 package dev.tanakornsss.luminality.data.backup
 
+import android.content.Context
+import android.net.Uri
 import com.google.gson.GsonBuilder
 import dev.tanakornsss.luminality.data.journal.JournalDao
 import dev.tanakornsss.luminality.data.model.Backup
@@ -9,11 +11,12 @@ import kotlinx.coroutines.withContext
 
 class BackupRepository(
     private val journalDao: JournalDao,
-    private val streakDao: StreakDao
+    private val streakDao: StreakDao,
+    private val context: Context
 ) {
     private val gson = GsonBuilder().setPrettyPrinting().create()
 
-    suspend fun exportJson(): String = withContext(Dispatchers.IO) {
+    suspend fun exportJson(uri: Uri) = withContext(Dispatchers.IO) {
         val journal = journalDao.exportAllEntries()
         val streak = streakDao.exportAllStreak()
 
@@ -22,10 +25,20 @@ class BackupRepository(
             streak = streak
         )
 
-        gson.toJson(data)
+        val res = gson.toJson(data)
+
+        context.contentResolver.openOutputStream(uri)?.use {
+            it.write(res.toByteArray())
+        }
     }
 
-    suspend fun importJson(json: String) = withContext(Dispatchers.IO) {
+    suspend fun importJson(uri: Uri) = withContext(Dispatchers.IO) {
+        val json = context.contentResolver.openInputStream(uri)?.use {
+            it.readBytes().decodeToString()
+        }
+
+        if (json == null) return@withContext
+
         val backup = gson.fromJson(json, Backup::class.java)
 
         journalDao.deleteAllEntries()
